Revert "repackaging completed. JNotebooks updated to use new package"
